In the "Create Fragment" flow, there should be a way to playback the fragment before posting it. Add a minimalist play icon (similar to what you have on the cards for each post on the "Home" page) to the bottom left of the fragment creation card, which will change to a pause icon when the fragment is currently playing in the same space.
While the fragment is playing, there should be a visual indicator that lights up each square in the 4x4 grid in sequence from left to right, top to bottom. To be clear, it should start with lighting up (with a visually pleasing outline on the square) the top left square (first square on the first row), proceed to the second square on the first row, then the third square on the first row, and so on.

The playback should not be at a random speed, instead, there should be a way to modify the BPM of the clip (similar to how you can modify the BPM in electronic music software). The default should be 120 BPM, which means there should be 120 beats per minute, or 2 beats per second. Each square in this grid should represent a beat, and the playback should be configured to be as fast as the BPM is set to.