We need to create a new API for posting a fragment. In the mused-api project we previously set up, can you create a new Firebase API for posting a fragment? The API should create a new document in the "fragments" collection in Firestore, which consists of the fields listed in the screenshot below. In the future, the author ID will pull the id from the user authentication, but for now, the author ID can be "0" and the author name can be "Staff" (similar to how the staff-uploaded sounds that we use in the sound marketplace have an author ID of "0"). Each fragment document in the "fragments" collection will have a number of rows and columns defining how long the fragment is. In the future, the number of rows and columns will be customizable, even though right now, the number of rows and columns in the grid is 4x4.
The "fragments" collection also needs to be extended so that it's possible to map each square to one or more sounds on that square. In the other screenshot below, the first pad (index 0) has the Roland 808 Kick Drum sound, the third pad (index 2) has 2 sounds, the Roland 808 Kick Drum sound and the Roland 808 Clap sound, etc. Since we want to create the functionality for remixing fragments in the future, we need to know which pads have which sounds assigned to them. This means that we'll need some data structure to store the relationship between pad position (row and column) and all the sound IDs associated with that pad. The sound IDs should be the same as the document IDs in the "sounds" collection in Firestore. Please let me know if you have any clarifying questions before you create the new post fragment API in the mused-api project