We need to create a new API for reading the list of current fragments. In the mused-api project we previously set up, can you create a new Firebase API for reading fragments? The API should look at the "fragments" collection in Firestore, which consists of the fields listed in the previous API you made for posting a new fragment. In the future, this API will pull a list of recommended fragments, including from artists you follow, but for now, this can be pulling from the full list of fragments in Firestore.
Using the mapping of pad index to sound ID (for each document in the "sounds" collection in Firestore), the relevant sound wav files should also be loaded into the Audio context. This is important because this read API will be used to create cards for each fragment on the Home page in the app. If the user clicks the "play" button at the top right of a fragment post, they should be able to hear what the fragment was, using the pad and sound mapping that's on the fragment document in Firestore. It should give them the same playback with the BPM from the document as if they were listening to it in the "Create Fragment" screen.

After you've created this API, can you update the client code to fetch the fragments using this API and preloading the sounds for playback so that the loaded fragments can play the same sounds that were assigned to that pad? Once that's done, add the response data to the existing mock fragment cards that are already on the home page, and reuse the same component structure for the real data from the API (referencing the sample post UI from the screenshot below):